Fix posts with multiple categories
Multiple categories in a post are treated differently in WordPress and
Jekyll. WordPress treats them as a list, while Jekyll treats them as a
hierarchy. This difference made some imported posts urls look odd,
like .../arduino/raspberry-pi/... as if raspberry-pi was a subcategory
of Arduino.

This change replaces the `categories` lists in all the posts with a
single `category` (more generic since there can only be one). For now
all legacy posts were changed to the `embedded` category. It also adds
a tag to a post that didn't have any (and with the category change, the
info would be lost).